chore(annotations): move descriptors into separate package
