+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/CommandExtractorImpl.java
@@ -23,6 +23,8 @@
 //
 package cloud.commandframework.annotations;
 
+import cloud.commandframework.annotations.descriptor.CommandDescriptor;
+import cloud.commandframework.annotations.descriptor.ImmutableCommandDescriptor;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
