+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/descriptor/FlagDescriptor.java
@@ -21,12 +21,14 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 //
-package cloud.commandframework.annotations;
+package cloud.commandframework.annotations.descriptor;
 
 import cloud.commandframework.Description;
 import cloud.commandframework.arguments.suggestion.SuggestionProvider;
 import cloud.commandframework.internal.ImmutableBuilder;
+import cloud.commandframework.permission.Permission;
 import java.lang.reflect.Parameter;
+import java.util.Collection;
 import org.apiguardian.api.API;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -35,15 +37,15 @@ import org.immutables.value.Value;
 @ImmutableBuilder
 @Value.Immutable
 @API(status = API.Status.STABLE, since = "2.0.0")
-public interface ArgumentDescriptor extends Descriptor {
+public interface FlagDescriptor extends Descriptor {
 
     /**
      * Creates a new builder.
      *
      * @return the created builder
      */
-    static ImmutableArgumentDescriptor.@NonNull Builder builder() {
-        return ImmutableArgumentDescriptor.builder();
+    static ImmutableFlagDescriptor.@NonNull Builder builder() {
+        return ImmutableFlagDescriptor.builder();
     }
 
     /**
@@ -54,12 +56,19 @@ public interface ArgumentDescriptor extends Descriptor {
     @NonNull Parameter parameter();
 
     /**
-     * Returns the argument name
+     * Returns the flag name.
      *
-     * @return the argument name
+     * @return the flag name
      */
     @Override
     @NonNull String name();
+
+    /**
+     * Returns an unmodifiable view of the flag aliases.
+     *
+     * @return the flag aliases
+     */
+    @NonNull Collection<@NonNull String> aliases();
 
     /**
      * Returns the name of the parser to use. If {@code null} the default parser for the parameter type will be used.
@@ -83,16 +92,23 @@ public interface ArgumentDescriptor extends Descriptor {
     @Nullable String suggestions();
 
     /**
-     * Returns the default value.
+     * Returns the permission of the flag.
      *
-     * @return the default value, or {@code null}
+     * @return the flag permission, or {@code null}
      */
-    @Nullable String defaultValue();
+    @Nullable Permission permission();
 
     /**
-     * Returns the description of the argument.
+     * Returns the description of the flag.
      *
-     * @return the argument description, or {@code null}
+     * @return the flag description, or {@code null}
      */
     @Nullable Description description();
+
+    /**
+     * Returns whether the flag is repeatable.
+     *
+     * @return whether the flag is repeatable
+     */
+    boolean repeatable();
 }
