+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/descriptor/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Objects that describe different command components.
+ */
+package cloud.commandframework.annotations.descriptor;
